# Change setTimeout to setImmediate