--- a/lib/private/intercept-exit-callbacks.js
+++ b/lib/private/intercept-exit-callbacks.js
@@ -167,9 +167,9 @@ module.exports = function interceptExitCallbacks (callbacks, _cache, hash, liveM
           if (liveMachine._runningSynchronously) {
             return proceed();
           }
-          setTimeout(function (){
+          setImmediate(function (){
             return proceed();
-          }, 0);
+          });
 
         })(function afterMaybeWaiting() {
 
@@ -244,7 +244,7 @@ module.exports = function interceptExitCallbacks (callbacks, _cache, hash, liveM
               // ===================================================================================================================
               if (liveMachine._doAlterStack) {
                 // Finally, copy the stack from `thisStack` created at the
-                // beginning of this function _before_ the setTimeout().
+                // beginning of this function _before_ the setImmediate().
 
                 // Set the new error's stack to that of `thisStack`, with the error message prepended.
                 // We prepend the message because the code below that modifies the stack expects