@mikermcneil 2016-10-19
@particlebanana I'd say so-- 23000% as fast.  LGTM-- i'll `npm link` and run tests for a few packs to try and catch any obvious problems that come from it, but I think we'll be fine.  Thanks!